+++ src/zepben/evolve/database/sqlite/writers/network_cim_writer.py
@@ -26,7 +26,8 @@ from zepben.evolve import CableInfo, TableCableInfo, PreparedStatement, WireInfo
     TableNoLoadTests, TableTransformerTest, TransformerTest, ShortCircuitTest, TableShortCircuitTests, OpenCircuitTest, TableOpenCircuitTests, \
     PerLengthImpedance, TablePerLengthImpedances, TableAssetOrganisationRoles, AssetOwner, TableAssetOwners, TableStructures, Structure, Pole, TablePoles, \
     Streetlight, TableStreetlights, Location, TableLocations, TableLocationStreetAddressField, StreetAddress, TableLocationStreetAddresses, PositionPoint, \
-    TablePositionPoints, TableStreetAddresses, TableTownDetails, TownDetail, TableEndDevices, Meter, TableMeters, TableUsagePoints, OperationalRestriction, \
+    TablePositionPoints, TableStreetAddresses, TableTownDetails, TownDetail, TableStreetDetails, StreetDetail, TableEndDevices, Meter, TableMeters, \
+    TableUsagePoints, OperationalRestriction, \
     TableOperationalRestrictions, TableFaultIndicators, TableAuxiliaryEquipment, AuxiliaryEquipment, FaultIndicator, TableMeasurements, Measurement, Analog, \
     TableAnalogs, Accumulator, TableAccumulators, Discrete, TableDiscretes, Control, TableControls, TableIoPoints, IoPoint, TableRemotePoints, RemotePoint, \
     RemoteControl, TableRemoteControls, RemoteSource, TableRemoteSources, ShuntCompensatorInfo, TableShuntCompensatorInfo, EquivalentBranch, \
@@ -209,6 +210,21 @@ class NetworkCIMWriter(BaseCIMWriter):
 
     # ************ IEC61968 COMMON ************
 
+    @staticmethod
+    def insert_street_detail(table: TableStreetDetails, insert: PreparedStatement, street_detail: StreetDetail):
+        insert.add_value(table.building_name.query_index, street_detail.building_name)
+        insert.add_value(table.floor_identification.query_index, street_detail.floor_identification)
+        insert.add_value(table.street_name.query_index, street_detail.name)
+        insert.add_value(table.number.query_index, street_detail.number)
+        insert.add_value(table.suite_number.query_index, street_detail.suite_number)
+        insert.add_value(table.type.query_index, street_detail.type)
+        insert.add_value(table.display_address.query_index, street_detail.display_address)
+
+    @staticmethod
+    def insert_town_detail(table: TableTownDetails, insert: PreparedStatement, town_detail: TownDetail):
+        insert.add_value(table.town_name.query_index, town_detail.name if town_detail else None)
+        insert.add_value(table.state_or_province.query_index, town_detail.state_or_province if town_detail else None)
+
     def save_location(self, location: Location) -> bool:
         table = self.database_tables.get_table(TableLocations)
         insert = self.database_tables.get_insert(TableLocations)
@@ -245,14 +261,12 @@ class NetworkCIMWriter(BaseCIMWriter):
     def save_street_address(self, table: TableStreetAddresses, insert: PreparedStatement, street_address: StreetAddress, street_id: str,
                             description: str) -> bool:
         insert.add_value(table.postal_code.query_index, street_address.postal_code)
+        insert.add_value(table.po_box.query_index, street_address.po_box)
 
-        return self.save_town_detail(table, insert, street_address.town_detail, street_id, description)
+        NetworkCIMWriter.insert_street_detail(table, insert, street_address.street_detail)
+        NetworkCIMWriter.insert_town_detail(table, insert, street_address.town_detail)
 
-    def save_town_detail(self, table: TableTownDetails, insert: PreparedStatement, town_detail: TownDetail, town_id: str, description: str) -> bool:
-        insert.add_value(table.town_name.query_index, town_detail.name if town_detail else None)
-        insert.add_value(table.state_or_province.query_index, town_detail.state_or_province if town_detail else None)
-
-        return self.try_execute_single_update(insert, town_id, description)
+        return self.try_execute_single_update(insert, street_id, description)
 
     # ************ IEC61968 METERING ************